% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestABBABA.R
\name{bestABBABA}
\alias{bestABBABA}
\title{Random Forest classifier}
\usage{
bestABBABA(data, test, file = FALSE)
}
\arguments{
\item{data}{BLABLABLA.}

\item{test}{BLABLABLA.}

\item{file}{Logical, set to TRUE if you want to create a csv file with the Id's and corresponding predicted classes, by deafult set to FALSE.}
}
\value{
A dataframe with following elements: Id, predicted class.
}
\description{
multi-class classifier that predicts the labels for the test data set using a random forest.
}
\examples{
# Given a training and a test set
bestABBABA(training, test)
}

