% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestABBABA.R
\name{bestABBABA.RF}
\alias{bestABBABA.RF}
\title{Random Forest classifier}
\usage{
bestABBABA.RF(data, test, file = FALSE, seed = 111)
}
\arguments{
\item{data}{Matrix or data frame consisting of the training data, where the labels are included in the matrix/data frame}

\item{test}{Matrix or data frame consisting of the test data.}

\item{file}{Logical, set to TRUE if you want to create a csv file with the Id's and corresponding predicted classes, by deafult set to FALSE.}

\item{seed}{Integer, set to 111 by default.}
}
\value{
A dataframe with following elements: Id, predicted class.
}
\description{
multi-class classifier that predicts the labels for the test data set using a random forest.
}
\examples{
#Run random forest
bestABBABA.RF(training, test)
}

